Outline inotify logic for /proc/ files
Outline the main plan on how to get notifications and
read/write permissions on kernel parameters.  As many
processes might have this ability, the following is decided:
- Subscribe to 'inotify' kernel events
- Update perceived content is file was written
- Track last modification timestamp

This possibly means, listening to inotify for a large number
of file paths, if recursive watch option is selected.
Handling of kernel events should be done by the GLib
main eventloop.